Refactor applyAltTextToMarkdown to shared MarkdownUtils
Moved the duplicated `applyAltTextToMarkdown` pure function from
`CreateVoiceActivity.kt` and `DashboardPostDetailActivity.kt` into the
shared `MarkdownUtils.kt` as `applyAltText`. This reduces code duplication
and improves maintainability while preserving existing behavior.
